+++ cloudfeaster/samples/run-sample.sh
@@ -78,8 +78,10 @@ if [ ! "${CLF_REMOTE_CHROMEDRIVER:-}" == "" ]; then NETWORK=host; fi
 
 DOCKER_CONTAINER_NAME=$(python -c "import uuid; print uuid.uuid4().hex")
 
-SPIDER_OUTPUT=$(mktemp 2> /dev/null || mktemp -t DAS)
-echo_if_verbose "Spider output @ '${SPIDER_OUTPUT}'"
+SPIDER_OUTPUT_ARTIFACT_DIR=$(mktemp -d 2> /dev/null || mktemp -d -t DAS)
+echo_if_verbose "Spider output artifact directory @ '${SPIDER_OUTPUT_ARTIFACT_DIR}'"
+
+SPIDER_OUTPUT=${SPIDER_OUTPUT_ARTIFACT_DIR}/spider-output.json
 
 docker run \
     --name "${DOCKER_CONTAINER_NAME}" \
@@ -91,12 +93,20 @@ docker run \
     "${DEV_ENV_DOCKER_IMAGE}" \
     "/app/$(repo.sh -u)/samples/${SPIDER}.py" "$@" >& "${SPIDER_OUTPUT}"
 
-SPIDER_OUTPUT_ARTIFACT_DIR=$(mktemp -d 2> /dev/null || mktemp -d -t DAS)
-echo_if_verbose "Spider output artifact directory @ '${SPIDER_OUTPUT_ARTIFACT_DIR}'"
+SPIDER_LOG_IN_CONTAINER=$(jq -r ._debug.spiderLog "${SPIDER_OUTPUT}" | sed -e "s|null||g")
+if [ ! "${SPIDER_LOG_IN_CONTAINER}" == "" ]; then
+    SPIDER_LOG_ON_HOST=${SPIDER_OUTPUT_ARTIFACT_DIR}/spider-log.txt
+
+    docker container cp \
+        "${DOCKER_CONTAINER_NAME}:${SPIDER_LOG_IN_CONTAINER}" \
+        "${SPIDER_LOG_ON_HOST}"
+
+    sed -i "" -e "s|${SPIDER_LOG_IN_CONTAINER}|${SPIDER_LOG_ON_HOST}|" "${SPIDER_OUTPUT}"
+fi
 
 CHROMEDRIVER_LOG_IN_CONTAINER=$(jq -r ._debug.chromeDriverLog "${SPIDER_OUTPUT}" | sed -e "s|null||g")
 if [ ! "${CHROMEDRIVER_LOG_IN_CONTAINER}" == "" ]; then
-    CHROMEDRIVER_LOG_ON_HOST=${SPIDER_OUTPUT_ARTIFACT_DIR}/$(basename "${CHROMEDRIVER_LOG_IN_CONTAINER}")
+    CHROMEDRIVER_LOG_ON_HOST=${SPIDER_OUTPUT_ARTIFACT_DIR}/chromedriver-log.txt
 
     docker container cp \
         "${DOCKER_CONTAINER_NAME}:${CHROMEDRIVER_LOG_IN_CONTAINER}" \